code addition to writeText() to check if the user gets writes on an existing file

+++ src/de/quinscape/Terminal.java
@@ -7,7 +7,6 @@ import java.util.UUID;
 
 public class Terminal {
     private static final Scanner scanner = new Scanner(System.in);
-    private boolean isOver = true;
 
     private File savedFile() {
         return new File("/Users/pchitanta/Documents/JavaCourseUdemy/Terminal/49e7b828-c9a2-46e4-93f5-" +
@@ -25,24 +24,36 @@ public class Terminal {
 
     private void writeText() {
         try {
-            java.io.FileWriter writer = new java.io.FileWriter(fileNameGenerator() + ".txt");
-            while (isOver) {
-                String userInput = scanner.nextLine();
-                if (!userInput.isEmpty()) {
-                    writer.append(userInput).append("\n");
-                } else {
-                    isOver = false;
-                    writer.close();
+            boolean isOver = true;
+            if (!doesFileExist(savedFile())) {
+                java.io.FileWriter writer = new java.io.FileWriter(fileNameGenerator() + ".txt");
+                while (isOver) {
+                    String userInput = scanner.nextLine();
+                    if (!userInput.isEmpty()) {
+                        writer.append(userInput).append("\n");
+                    } else {
+                        isOver = false;
+                        writer.close();
+                    }
+                }
+            } else {
+                while (isOver) {
+                    String userInput = scanner.nextLine();
+                    try {
+                        java.io.FileWriter writer = new java.io.FileWriter(savedFile());
+                        if (!userInput.isEmpty()) {
+                            writer.append(userInput).append("\n");
+                        } else {
+                            isOver = false;
+                            writer.close();
+                        }
+                    } catch (IOException e) {
+                        System.out.println("Error: " + e.getMessage());
+                    }
                 }
             }
         } catch (IOException e) {
             System.out.println("Error: " + e.getMessage());
-        }
-    }
-
-    private void appendText() throws IOException {
-        try (java.io.FileWriter wr = new java.io.FileWriter("49e7b828-c9a2-46e4-93f5-cbecf45a013c.txt", true)) {
-            wr.append("\n").append(scanner.nextLine());
         }
     }
 
@@ -62,16 +73,10 @@ public class Terminal {
         }
     }
 
-    public void startApplication() throws IOException {
-        if (doesFileExist(savedFile())) {
-            printCheckIfFileExists();
-            printInstructions();
-            appendText();
-        } else {
-            printCheckIfFileExists();
-            printInstructions();
-            writeText();
-        }
+    public void startApplication() {
+        printCheckIfFileExists();
+        printInstructions();
+        writeText();
     }
 }
 //TODO ich will das mein UUID nur 8 Zeichen erstellt
